+++ lib/src/health.dart
@@ -4,16 +4,21 @@
 
 library pana.health;
 
+import 'package:meta/meta.dart';
+
 import 'messages.dart' as messages;
 import 'model.dart';
 import 'pubspec.dart';
 
 /// Extracts and summarizes the output of `dartanalyzer` and `dartfmt`.
-Health calcHealth(
-  Pubspec pubspec,
-  List<CodeProblem> analyzerItems,
-  Iterable<DartFileSummary> dartFileSummaries,
-) {
+Health calcHealth({
+  @required Pubspec pubspec,
+  @required bool analyzeProcessFailed,
+  @required bool formatProcessFailed,
+  @required bool resolveProcessFailed,
+  @required List<CodeProblem> analyzerItems,
+  @required Iterable<DartFileSummary> dartFileSummaries,
+}) {
   analyzerItems ??= const <CodeProblem>[];
   final analyzerErrorCount = analyzerItems.where((c) => c.isError).length;
   final analyzerWarningCount = analyzerItems.where((c) => c.isWarning).length;
@@ -88,6 +93,9 @@ Health calcHealth(
   suggestions.sort();
 
   return new Health(
+    analyzeProcessFailed: analyzeProcessFailed,
+    formatProcessFailed: formatProcessFailed,
+    resolveProcessFailed: resolveProcessFailed,
     analyzerErrorCount: analyzerErrorCount,
     analyzerWarningCount: analyzerWarningCount,
     analyzerHintCount: analyzerHintCount,