+++ lib/src/package_analyzer.dart
@@ -328,7 +328,14 @@ class PackageAnalyzer {
       );
     }
 
-    final health = calcHealth(pubspec, analyzerItems, files.values);
+    final health = calcHealth(
+      pubspec: pubspec,
+      analyzeProcessFailed: analyzerItems == null,
+      formatProcessFailed: unformattedFiles == null,
+      resolveProcessFailed: pkgResolution == null,
+      analyzerItems: analyzerItems,
+      dartFileSummaries: files.values,
+    );
 
     DartPlatform platform;
     if (pkgPlatformConflict != null) {